requirements.txt
anytree==2.8.0
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.6
django-admin-csvexport==1.8
django-cors-headers==3.3.0
django-csv-exports==1.0.5
django-richtextfield==1.6
djangorestframework==3.11.0
drf-action-serializer==1.0.5
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.9
inflection==0.4.0
install==1.3.4
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
packaging==20.4
pandas==1.1.5
Pillow==7.1.2
psycopg2==2.8.5
pyparsing==2.4.7
python-dateutil==2.8.1
python-http-client==3.3.1
pytz==2020.1
requests==2.23.0
rest-condition==1.0.3
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
sendgrid==6.4.8
six==1.15.0
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.1.0
